Time: 0 ms (100%), Space: 40.8 MB (89.71%) - LeetHub

+++ 3201-distribute-candies-among-children-ii/3201-distribute-candies-among-children-ii.java
@@ -1,11 +1,14 @@
 class Solution {
     public long distributeCandies(int n, int limit) {
+        //Use stars and bars method from permutation and combination
+
+
         return combination(n) - 3L * combination(n - (limit + 1)) + 3L * combination(n - 2 * (limit + 1)) - combination(n - 3 * (limit + 1));
     }
 
-    public long combination(int a){
+    public long combination(long a){
         if(a < 0)
             return 0;
-        return ((long)(a + 1) * (a + 2)) / 2L;
+        return ((a + 1) * (a + 2)) / 2L;
     }
 }
